requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.24.4
scikit-learn==1.3.2
networkx==3.2.1
scipy==1.11.4
plotly==5.17.0
dash==2.14.2
redis==5.0.1
python-multipart==0.0.6
pydantic==2.5.0
kaggle==1.5.16
seaborn==0.12.2
jupyter==1.0.0
matplotlib==3.7.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
